add auto 42 profile picture import and fixes about empty picture form

+++ pong_back/users/models.py
@@ -1,5 +1,8 @@
 from django.db import models
 from django.contrib.auth.models import User
+from django.core.files import File
+from django.core.files.uploadedfile import InMemoryUploadedFile
+from django.core.files.temp import NamedTemporaryFile
 from django.utils.crypto import get_random_string
 from django.contrib.auth import update_session_auth_hash, login, authenticate
 from django.http import HttpResponse, HttpRequest
@@ -10,13 +13,16 @@ from uuid import uuid4
 from datetime import timedelta
 from stats.models import Stats
 import os
+import requests
 
 # instance corresponding to the instanse of imagefield automaticcly passed
-def generateUniqueImageID(instance, filename):
+def generateUniqueImageID(instance, filename, extension=None):
+	print(f"le filename {filename}")
 	if (filename == "pokemon.png"):
 		return (filename)
 	
-	extension = filename.split('.')[-1]
+	if not extension:
+		extension = filename.split('.')[-1]
 	uniqueId = str(uuid4())
 	finalPath = f'{uniqueId}.{extension}'
 
@@ -30,8 +36,7 @@ class Profile(models.Model):
 	user: User = models.OneToOneField(User, on_delete=models.CASCADE, blank=False, primary_key=True)
 	friends = models.ManyToManyField(User, blank=True, symmetrical=True, related_name="friends")
 	pendingFriendsFrom = models.ManyToManyField(User, related_name="pendingFriendsFrom", symmetrical=False, blank=True)
-	is_2fa = models.BooleanField(default=False)
-	profilePicture = models.ImageField(upload_to=generateUniqueImageID, default="default_e04bed5e-0be0-441b-9616-41f09b84aaf7.png", blank=True) #have to be resize
+	profilePicture = models.ImageField(upload_to=generateUniqueImageID, default=settings.DEFAULT_PROFILE_PICTURE_NAME, blank=True)
 	blockedUsers = models.ManyToManyField(User, related_name="blockedUsers", symmetrical=False, blank=True)
 	lastPasswordChange = models.DateTimeField(default=now, blank=True)
 	
@@ -60,7 +65,6 @@ class Profile(models.Model):
 		else:
 			private = {
 				"email": self.user.email,
-				"is_2fa": self.is_2fa,
 				"friends": self.getManyToTab(self.friends),
 				"pendingFriendsFrom": self.getManyToTab(self.pendingFriendsFrom),
 				"blockedUsers": self.getManyToTab(self.blockedUsers)
@@ -109,17 +113,19 @@ class Profile(models.Model):
 		from .forms import PictureForm
 		
 		oldPicture = self.profilePicture.path
-		form: PictureForm = PictureForm(request.POST, request.FILES, instance=self)
+		form: PictureForm = PictureForm(request.POST, request.FILES)
 
 		if (form.is_valid()):
-			if (os.path.exists(oldPicture)):
+			if (os.path.exists(oldPicture) and os.path.basename(oldPicture) != settings.DEFAULT_PROFILE_PICTURE_NAME):
 				os.remove(oldPicture)
 
-			form.save()
+			self.profilePicture = form.cleaned_data['profilePicture']
+			self.save()
 			return (tResponses.OKAY.request(message="Profile picture successfully changed !"))
 		else:
 			return (tResponses.BAD_REQUEST.request(message="Image is not valid !"))
 	
+
 	# Will check if the passed user is blocked
 	def is_block(self, target: User):
 		if target in self.blockedUsers.all():
@@ -152,14 +158,27 @@ class Profile(models.Model):
 	
 	# return 0 if success then 1 
 	@staticmethod
-	def login42(request: HttpRequest, username: str, email: str):
+	def login42(request: HttpRequest, username: str, email: str, image: str):
 		username = f'42_{username}'
 		existing_account = Profile.getUserFromUsername(username)
 
 		if not existing_account: #mean that we want to login
 			existing_account = Profile.registerUser(username, email)
 			if not existing_account :
-				return 1			
+				return 1
+			# Now retrieving 42 Profile picture
+			response = requests.get(image)
+			if response.status_code == 200:
+				img_temp = NamedTemporaryFile(delete=True)
+				img_temp.write(response.content)
+				img_temp.flush()
+				print(response.content)
+				img = InMemoryUploadedFile(img_temp, None, generateUniqueImageID(None, img_temp.name, image.split('.')[-1]), 'image/jpeg', len(response.content), None)
+
+				existing_account.profile.profilePicture = img
+				existing_account.profile.save()
+
+				img_temp.close()
 		
 		Profile.createUserOnetoOne(existing_account)
 		login(request, existing_account)
